requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
nltk==3.8.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
transformers==4.35.0
sentencepiece==0.1.99
torch==2.1.0
accelerate==0.24.1
mediapipe==0.10.9
